move ensure_stream_ingestion_jobs back out of the client
